requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
gunicorn==23.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
firebase-admin==6.6.0
google-auth==2.36.0
httpx==0.27.2
redis==5.1.1
rq==2.1.0
python-dotenv==1.0.1
yt-dlp==2025.1.26
requests==2.32.3
google-cloud-firestore==2.20.0
